Revert "Admin API and schema"
